pull/21705  Fix COPP ARP Test
The COPP test for the ARP packet type does not
account for the fact that there are two ARP
trap IDs being acted upon in the target COPP trap
group: arp request and arp response.

Set the dst_ip to a non-reachable IP to avoid
generating an ARP response which would, depending
on the underlying trap rate calculation being
used, erroneously yield double the nominal rate.

Signed-off-by: Christopher Croy <ccroy@arista.com>

Improve RX PPS Calcuation for COPP Tests

The current calculation method for the RX PPS rate
for the COPP tests is not very accurate (in some
cases 130-150% above nominal) due to the reliance
of sampling received packets at the ptf container
on the testbed server after sending the packet
stream, while also using a timeout to wait for
said packets to finish arriving. This does not
capture the in-flight RX PPS rate, but rather
takes an average outside of the actual packet
transmission window, and incurs additional
inaccuracies due to the wait time at the end.

A more accurate approach implemented here is to
take two snapshots of the RX packet count at the
NN agent on the dut itself while the packet stream
is already running.

Signed-off-by: Christopher Croy <ccroy@arista.com>